refactor(webclipper): move backup into src/sync/backup
