+++ Extensions/WebClipper/src/sync/backup/export.ts
@@ -1,13 +1,13 @@
-import { storageGetAll } from '../platform/storage/local';
+import { storageGetAll } from '../../platform/storage/local';
 import {
   BACKUP_ZIP_SCHEMA_VERSION,
   filterStorageForBackup,
   uniqueConversationKey,
 } from './backup-utils';
-import { buildConversationBasename } from '../conversations/file-naming';
+import { buildConversationBasename } from '../../conversations/file-naming';
 import { openDb, reqToPromise, tx, txDone } from './idb';
 import { createZipBlob } from './zip-utils';
-import runtimeContext from '../runtime-context.ts';
+import runtimeContext from '../../runtime-context.ts';
 
 type AnyRecord = Record<string, any>;
