+++ Extensions/WebClipper/src/sync/backup/idb.ts
@@ -1,4 +1,4 @@
-import { openDb as openSchemaDb } from '../platform/idb/schema';
+import { openDb as openSchemaDb } from '../../platform/idb/schema';
 
 let cachedDb: IDBDatabase | null = null;
 let openingDb: Promise<IDBDatabase> | null = null;
