+++ Extensions/WebClipper/src/sync/local/zip-utils.ts
@@ -1,9 +1,5 @@
 import runtimeContext from '../../runtime-context.ts';
-import {
-  createZipBlob,
-  extractZipEntries,
-  isUnsafeZipEntryName,
-} from '../../backup/zip-utils';
+import { createZipBlob, extractZipEntries, isUnsafeZipEntryName } from '../backup/zip-utils';
 
 const api = { createZipBlob, extractZipEntries, isUnsafeZipEntryName };
 (runtimeContext as any).zipUtils = api;
