+++ Extensions/WebClipper/tests/domains/backup-zip-utils.test.ts
@@ -1,7 +1,7 @@
 import { deflateRawSync } from 'node:zlib';
 import { describe, expect, it } from 'vitest';
 
-import { createZipBlob, extractZipEntries } from '../../src/backup/zip-utils';
+import { createZipBlob, extractZipEntries } from '../../src/sync/backup/zip-utils';
 
 function u16(n: number) {
   return new Uint8Array([n & 0xff, (n >>> 8) & 0xff]);
